GĐ 5-1: Xây dựng giao diện quản lý task.

+++ routes/web/tasks.php
@@ -8,22 +8,20 @@ Route::middleware(['auth', 'verified'])
     ->controller(TaskController::class)
     ->group(function () {
         Route::get('/', 'index')->name('index');
-        Route::get('{task}', 'show')->name('show');
-
         Route::middleware('role:admin|super-admin|manager|leader')->group(function () {
             Route::get('create', 'create')->name('create');
             Route::post('/', 'store')->name('store');
 
-            Route::middleware('role:admin|super-admin|manager')->group(function () {
-                Route::get('recycle', 'recycle')->name('recycle');
-                Route::delete('{task}', 'softDelete')->name('soft-delete');
-                Route::post('{task}/restore', 'restore')->name('restore');
-                Route::delete('{task}/force-delete', 'forceDelete')->name('force-delete');
-            });
+            Route::get('recycle', 'recycle')->name('recycle');
+            Route::delete('{task}', 'softDelete')->name('soft-delete');
+            Route::post('{task}/restore', 'restore')->withTrashed()->name('restore');
+
+            Route::delete('{task}/force-delete', 'forceDelete')->withTrashed()->middleware('role:admin|super-admin', 'password.confirm')->name('force-delete');
         });
 
         Route::middleware('role:admin|super-admin|manager|leader|member')->group(function () {
-            Route::get('{task}/edit', 'edit')->name('edit');
             Route::put('{task}', 'update')->name('update');
         });
+
+        Route::get('{task}', 'show')->name('show');
     });
